backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Socket.IO - CRITICAL: Keep version 5.10.0
python-socketio==5.10.0
python-engineio==4.9.0

# Configuration & Settings
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# LangChain & AI
langgraph==0.0.32
langchain==0.1.6
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Vector DB
qdrant-client==1.7.3

# Speech Services
deepgram-sdk==3.2.0
elevenlabs==0.2.27

# Speech Recognition
openai-whisper==20231117
faster-whisper==0.10.0

# Document Processing
docling==2.0.1
docling-core==1.0.0

# Logging
python-json-logger==2.0.7

# WebSockets
websockets==12.0
